move frome mixture to mixtureS
